+++ lecture01/src/main/java/io/github/javaasasecondlanguage/lecture01/practice2/Palindrome.java
@@ -1,4 +1,4 @@
-package io.github.javaasasecondlanguage.lecture01;
+package io.github.javaasasecondlanguage.lecture01.practice2;
 
 public class Palindrome {
     /**
