+++ lecture01/src/test/java/io/github/javaasasecondlanguage/lecture01/practice1/HelloWorldTest.java
@@ -1,4 +1,4 @@
-package io.github.javaasasecondlanguage.lecture01;
+package io.github.javaasasecondlanguage.lecture01.practice1;
 
 import org.junit.jupiter.api.Test;
