+++ lecture01/src/test/java/io/github/javaasasecondlanguage/lecture01/practice2/PalindromeTest.java
@@ -1,4 +1,4 @@
-package io.github.javaasasecondlanguage.lecture01;
+package io.github.javaasasecondlanguage.lecture01.practice2;
 
 import org.junit.jupiter.api.Test;
